requirements.txt
fastapi==0.112.0
sqlalchemy==2.0.31
alembic==1.13.2
pyjwt==2.9.0
passlib==1.7.4
bcrypt==4.0.1
pydantic_settings==2.4.0
email_validator==2.2.0